Refactor registration handling by introducing openRegistrationEditor and registerMidashigo function.
