+++ src/extension.ts
@@ -4,6 +4,7 @@ import * as vscode from 'vscode';
 import { IInputMode } from './input-mode/IInputMode';
 import { AsciiMode } from './input-mode/AsciiMode';
 import * as jisyo from './jisyo/jisyo';
+import { registerMidashigo } from './input-mode/henkan/RegistrationEditor';
 
 var timestampOfCursorMoveCausedByKeyInput: number | undefined = undefined;
 
@@ -125,57 +126,8 @@ export async function activate(context: vscode.ExtensionContext) {
 	});
 	context.subscriptions.push(symbolInput);
 
-	const registerCandidateCommand = vscode.commands.registerCommand('skk.registerCandidate', async () => {
-		/**
-		 * 現在のエディタの内容が下の形式にあてはまっている場合に、その内容をユーザ辞書に登録する。
-		 * 読みが「あt」で単語が「合」の場合:
-		 * 読み:あt
-		 * 単語:合
-		 * 
-		 * ユーザ辞書のキー「あt」に対して，値「合」を先頭に追加する。
-		 * その後，エディタを閉じる。
-		 */
-		const editor = vscode.window.activeTextEditor;
-		if (!editor) {
-			return;
-		}
-		const document = editor.document;
-		const content = document.getText();
-		const lines = content.split('\n');
-
-		// フォーマットの確認
-		if (lines.length < 2 // 2行以上でなければならない
-			|| lines.slice(2).some(line => line.trim() !== '') // 3行目以降は空でなければならない
-			|| lines[0].slice(0, 3) !== "読み:" // 1行目は「読み:」で始まらなければならない
-			|| lines[1].slice(0, 3) !== "単語:" // 2行目は「単語:」で始まらなければならない
-		) {
-			vscode.window.showErrorMessage("SKK: 辞書登録できません。フォーマットが不正です。");
-			return;
-		}
-
-		const yomi = lines[0].slice(3);
-		const word = lines[1].slice(3);
-
-		if (yomi === '') {
-			vscode.window.showErrorMessage("SKK: 辞書登録できません。読みが空です。");
-			return;
-		}
-		if (word === '') {
-			vscode.window.showErrorMessage("SKK: 辞書登録できません。単語が空です。");
-			return;
-		}
-
-		const newCandidateList = [{ word, annotation: undefined }, ...jisyo.getGlobalJisyo().get(yomi) || []];
-		// dedup
-		const deduped = newCandidateList.filter((candidate, index, self) => self.findIndex(c => c.word === candidate.word) === index);
-		jisyo.getGlobalJisyo().set(yomi, deduped);
-
-		// clear all text in the editor
-		await editor.edit(editBuilder => {
-			editBuilder.delete(new vscode.Range(new vscode.Position(0, 0), new vscode.Position(document.lineCount, 0)));
-		});
-		// close the editor
-		await vscode.commands.executeCommand('workbench.action.closeActiveEditor');
+	const registerCandidateCommand = vscode.commands.registerCommand('skk.registerMidashigo', async () => {
+		await registerMidashigo();
 	});
 	context.subscriptions.push(registerCandidateCommand);
